Move functions out of views modules into function.py and simplify
