+++ networkx/classes/__init__.py
@@ -2,10 +2,10 @@ from .graph import Graph
 from .digraph import DiGraph
 from .multigraph import MultiGraph
 from .multidigraph import MultiDiGraph
-from .reportviews import *
-from .graphviews import *
 from .ordered import *
 
 from .function import *
 
-import coreviews
+import networkx.classes.coreviews
+import networkx.classes.graphviews
+import networkx.classes.reportviews
