+++ networkx/classes/function.py
@@ -25,6 +25,8 @@ from networkx.utils import pairwise, not_implemented_for
 __all__ = ['nodes', 'edges', 'degree', 'degree_histogram', 'neighbors',
            'number_of_nodes', 'number_of_edges', 'density',
            'is_directed', 'info', 'freeze', 'is_frozen', 'subgraph',
+           'induced_subgraph', 'edge_subgraph', 'reverse_view',
+           'to_directed', 'to_undirected',
            'add_star', 'add_path', 'add_cycle',
            'create_empty_copy', 'set_node_attributes',
            'get_node_attributes', 'set_edge_attributes',
@@ -302,6 +304,124 @@ def subgraph(G, nbunch):
     subgraph(G) calls G.subgraph()
     """
     return G.subgraph(nbunch)
+
+
+def induced_subgraph(G, nbunch):
+    """Return a SubGraph view of `G` showing only nodes in nbunch.
+
+    The induced subgraph of a graph on a set of nodes N is the
+    graph with nodes N and edges from G which have both ends in N.
+
+    Parameters
+    ----------
+    G : NetworkX Graph
+    nbunch : node, container of nodes or None (for all nodes)
+
+    Returns
+    -------
+    subgraph : SubGraph View
+        A read-only view of the subgraph in `G` induced by the nodes.
+        Changes to the graph `G` will be reflected in the view.
+
+    Notes
+    -----
+    To create a mutable subgraph with its own copies of nodes
+    edges and attributes use `subgraph.copy()` or `Graph(subgraph)`
+
+    For an inplace reduction of a graph to a subgraph you can remove nodes:
+    `G.remove_nodes_from(n in G if n not in set(nbunch))`
+
+    Examples
+    --------
+    >>> import networkx as nx
+    >>> G = nx.path_graph(4)  # or DiGraph, MultiGraph, MultiDiGraph, etc
+    >>> H = G.subgraph([0, 1, 2])
+    >>> list(H.edges)
+    [(0, 1), (1, 2)]
+    """
+    induced_nodes = nx.filters.show_nodes(G.nbunch_iter(nbunch))
+    if G.is_multigraph():
+        if G.is_directed():
+            return nx.graphviews.MultiDiSubGraph(G, induced_nodes)
+        return nx.graphviews.MultiSubGraph(G, induced_nodes)
+    if G.is_directed():
+        return nx.graphviews.DiSubGraph(G, induced_nodes)
+    return nx.graphviews.SubGraph(G, induced_nodes)
+
+
+def edge_subgraph(G, edges):
+    """Returns a view of the subgraph induced by the specified edges.
+
+    The induced subgraph contains each edge in `edges` and each
+    node incident to any of those edges.
+
+    Parameters
+    ----------
+    G : NetworkX Graph
+    edges : iterable
+        An iterable of edges. Edges not present in `G` are ignored.
+
+    Returns
+    -------
+    subgraph : SubGraph View
+        A read-only edge-induced subgraph of `G`.
+        Changes to `G` are reflected in the view.
+
+    Notes
+    -----
+    To create a mutable subgraph with its own copies of nodes
+    edges and attributes use `subgraph.copy()` or `Graph(subgraph)`
+
+    Examples
+    --------
+    >>> import networkx as nx
+    >>> G = nx.path_graph(5)
+    >>> H = G.edge_subgraph([(0, 1), (3, 4)])
+    >>> list(H.nodes)
+    [0, 1, 3, 4]
+    >>> list(H.edges)
+    [(0, 1), (3, 4)]
+    """
+    edges = set(edges)
+    nodes = set()
+    for e in edges:
+        nodes.update(e[:2])
+    induced_edges = nx.classes.filters.show_edges(edges)
+    induced_nodes = nx.classes.filters.show_nodes(nodes)
+    if G.is_multigraph():
+        if G.is_directed():
+            return nx.graphviews.MultiDiSubGraph(G, induced_nodes, induced_edges)
+        return nx.graphviews.MultiSubGraph(G, induced_nodes, induced_edges)
+    if G.is_directed():
+        return nx.graphviews.DiSubGraph(G, induced_nodes, induced_edges)
+    return nx.graphviews.SubGraph(G, induced_nodes, induced_edges)
+
+
+@not_implemented_for('undirected')
+def reverse_view(digraph):
+    if digraph.is_multigraph():
+        return nx.graphviews.MultiReverseView(digraph)
+    return nx.graphviews.ReverseView(digraph)
+
+
+def to_directed(graph):
+    if graph.is_directed():
+        if graph.is_multigraph():
+            return induced_subgraph(graph, graph)
+        return induced_subgraph(graph, graph)
+    if graph.is_multigraph():
+        return nx.graphviews.MultiDirectedView(graph)
+    return nx.graphviews.DirectedView(graph)
+
+
+def to_undirected(digraph):
+    if digraph.is_directed():
+        if digraph.is_multigraph():
+            return nx.graphviews.MultiUnDirectedView(digraph)
+        return nx.graphviews.UnDirectedView(digraph)
+    if digraph.is_multigraph():
+        return induced_subgraph(digraph, digraph)
+    return induced_subgraph(digraph, digraph)
 
 
 def create_empty_copy(G, with_data=True):
